+++ components/checks/passive/grep/insecure_cors_policy.rb
@@ -7,16 +7,18 @@
 =end
 
 # @author  Tasos Laskos <tasos.laskos@arachni-scanner.com>
-# @version 0.1
+# @version 0.1.1
 class Arachni::Checks::InsecureCORSPolicy < Arachni::Check::Base
 
     def run
         return if audited?( page.parsed_url.host ) ||
             page.response.headers['Access-Control-Allow-Origin'] != '*'
-
         audited( page.parsed_url.host )
 
-        log( vector: Element::Server.new( page.url ) )
+        log(
+            vector: Element::Server.new( page.url ),
+            proof:  page.response.headers_string[/Access-Control-Allow-Origin.*$/i]
+        )
     end
 
     def self.info
@@ -24,7 +26,7 @@ class Arachni::Checks::InsecureCORSPolicy < Arachni::Check::Base
             name:        'Insecure CORS policy',
             description: %q{Checks the host for a wildcard (`*`) `Access-Control-Allow-Origin` header.},
             author:      'Tasos Laskos <tasos.laskos@arachni-scanner.com>',
-            version:     '0.1',
+            version:     '0.1.1',
             elements:    [ Element::Server ],
 
             issue:       {